feat(fitted): markdown resume editor on dedicated route
Replaces the modal ResumeEditor with a full-page editor at
/fitted/jobs/[id]/resume. The editor takes markdown directly (matching
the new payload_md source-of-truth), surfaces lint warnings, and renders
metadata + version history. Approve navigates back so the parent
re-fetches the updated status.
