+++ apps/root/src/app/fitted/(app)/jobs/[id]/resume/page.tsx
@@ -1,0 +1,15 @@
+import type { Metadata } from 'next';
+import ResumeReviewPage from './ResumeReviewPage';
+
+export const metadata: Metadata = {
+  title: 'Review Resume',
+};
+
+export default async function FittedResumeReview({
+  params,
+}: {
+  params: Promise<{ id: string }>;
+}) {
+  const { id } = await params;
+  return <ResumeReviewPage jobPostingId={id} />;
+}
